requirements.txt
psycopg2
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt